Separate ResizeObserver from scroller
